Add files via upload
Objective: Analyze the flight booking dataset obtained from the “Ease My Trip” website, as well as the dataset obtained from Kaggle

Dataset: Contains information about flight booking options for travel between India's top 6 metro cities.

Data points: 300261 data points in total. There are 11 features in the dataset.

Importance: A thorough analysis of the data will aid in the discovery of valuable insights. These insights will be of enormous value to passengers.

GOALS:
Need to understand the dataset and Analyse the throughout dataset

Our aim of this analysis is to answer the below research questions:
	a) Does price vary with Airlines?	b) How is the price affected when tickets are bought in just 1 or 2 days before departure?	c) Does ticket price change based on the departure time and arrival time?	d) How the price changes with change in Source and Destination?	e) How does the ticket price vary between Economy and Business class?

Construct and train a regression model. Using this model to predict the prices of future tickets

Determine which factors have the greatest impact on price by analysing the trained data